requirements.txt
aiogram==3.18.0
asyncpg==0.30.0
fastapi==0.115.8
greenlet==3.1.1
httpx==0.28.1
Pillow==11.1.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
redis==5.2.1
SQLAlchemy==2.0.38
uvicorn==0.34.0
pytest==8.3.4
pytest-asyncio==0.25.3
